+++ src/app/shared/NoteContentEllipsis.svelte
@@ -2,6 +2,6 @@
   import Anchor from "src/partials/Anchor.svelte"
 </script>
 
-<div class="relative pointer-events-none -mt-4 mb-0 flex justify-center z-10">
-  <Anchor button low>See more</Anchor>
+<div class="relative -mt-4 mb-0 flex justify-center z-10">
+  <Anchor stopPropagation button low on:click>See more</Anchor>
 </div>